+++ lib/src/validator/all_validations.dart
@@ -139,6 +139,7 @@ class AllValidations {
 
   /// Valida se o número é um telefone fixo brasileiro com DDD.
   static bool isBrazilianLandline(String s) {
+    if (s.isEmpty) return false;
     // Remove caracteres não numéricos
     String cleanedNumber = removeCharacters(s);
 
@@ -163,77 +164,7 @@ class AllValidations {
   }
 
   static bool isValidDDD(String ddd) {
-    const validDDDs = [
-      "11",
-      "12",
-      "13",
-      "14",
-      "15",
-      "16",
-      "17",
-      "18",
-      "19",
-      "21",
-      "22",
-      "24",
-      "27",
-      "28",
-      "31",
-      "32",
-      "33",
-      "34",
-      "35",
-      "37",
-      "38",
-      "41",
-      "42",
-      "43",
-      "44",
-      "45",
-      "46",
-      "47",
-      "48",
-      "49",
-      "51",
-      "53",
-      "54",
-      "55",
-      "61",
-      "62",
-      "63",
-      "64",
-      "65",
-      "66",
-      "67",
-      "68",
-      "69",
-      "71",
-      "73",
-      "74",
-      "75",
-      "77",
-      "79",
-      "81",
-      "82",
-      "83",
-      "84",
-      "85",
-      "86",
-      "87",
-      "88",
-      "89",
-      "91",
-      "92",
-      "93",
-      "94",
-      "95",
-      "96",
-      "97",
-      "98",
-      "99"
-    ];
-
-    return validDDDs.contains(ddd);
+    return Constants.ddds.contains(ddd);
   }
 
   /// Retorna o estado correspondente ao DDD informado.
@@ -582,6 +513,12 @@ class AllValidations {
       }
     }
     return true;
+  }
+
+  static bool isValidBrazilianLicensePlate(String plate) {
+    final oldModel = RegExp(r'^[A-Z]{3}-?\d{4}$'); // Ex.: ABC-1234
+    final newModel = RegExp(r'^[A-Z]{3}\d[A-Z]\d{2}$'); // Ex.: ABC1D23
+    return oldModel.hasMatch(plate) || newModel.hasMatch(plate);
   }
 
   ///check if password is equal to confirm password or pharse is equal to confirm phrase