+++ test/all_validations_br_test.dart
@@ -329,6 +329,12 @@ void main() {
       expect(AllValidations.isValidDDD("80"), false); // Não existe DDD 80
     });
   });
+
+   test('isValidBrazilianLicensePlate', () {
+      expect(AllValidations.isValidBrazilianLicensePlate('ABC-1234'), isTrue);
+      expect(AllValidations.isValidBrazilianLicensePlate('ABC1D23'), isTrue);
+      expect(AllValidations.isValidBrazilianLicensePlate('INVALID'), isFalse);
+    });
   group('Validação de Celulares Brasileiros', () {
     test('Números de celular válidos', () {
       expect(AllValidations.isBrazilianCellPhone("(11) 91234-5678"), true);